refactor&feature: Organize directories and Implement save relation entity
