+++ server/src/auth/types/request.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/users/users.entity';
+import { User } from 'src/users/entities/users.entity';
 
 export type RequestWithUser = Request & {
   user: User;
