+++ server/src/teams/teams.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersModule } from 'src/users/users.module';
-import { Team } from './teams.entity';
+import { Team } from './entities/teams.entity';
 import { TeamsResolver } from './teams.resolver';
 import { TeamsService } from './teams.service';
