+++ server/src/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { UsersService } from './users.service';
-import { User } from './users.entity';
+import { User } from './entities/users.entity';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UsersResolver } from './users.resolver';
